Rename WndMainWindow to MainWindow
