+++ Windows/GEDebugger/GEDebugger.cpp
@@ -29,7 +29,7 @@
 #include "Windows/GEDebugger/TabVertices.h"
 #include "Windows/InputBox.h"
 #include "Windows/WindowsHost.h"
-#include "Windows/WndMainWindow.h"
+#include "Windows/MainWindow.h"
 #include "Windows/main.h"
 #include "GPU/GPUInterface.h"
 #include "GPU/Common/GPUDebugInterface.h"
